Quick attempt at adding 'delete last statement' (#40)
The most recent stando statement can now be deleted from the database by using the command
```
!stando That's been debunked
```

Changed the way we interact with the database to accomodate this, rather than keeping the list of statements in memory, we now query the database for a random statement every time the `!stando` command is run. This may be a bit slower, but it let's us update the database externally while the bot is running which is handy.